requirements.txt
Flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.1.1
youtube-transcript-api==1.2.4
yt-dlp==2026.3.31
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
faiss-cpu==1.11.0
openai==1.75.0

